requirements.txt
beautifulsoup4==4.6.0
python-decouple==3.1
requests==2.18.4